Add feature flag to enable/disable registration journey, and rename things so it makes more sense
