+++ packages/webapp/.jest/test.env.js
@@ -19,4 +19,5 @@ process.env.ENABLE_ROUTE_SUPPORT_FOR_GEOSPATIAL='Y'
 process.env.ENABLE_ROUTE_SUPPORT_FOR_ADDITIONAL_EMAIL='Y'
 process.env.ENABLE_ROUTE_SUPPORT_FOR_DEV_JOURNEY='Y'
 process.env.ENABLE_ROUTE_SUPPORT_FOR_CREDIT_ESTIMATION_JOURNEY='Y'
+process.env.ENABLE_ROUTE_SUPPORT_FOR_REGISTRATION_JOURNEY='Y'
 process.env.DEFRA_ID_MOCK=false
